@@ -1,222 +1,240 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>au.com.mineauz</groupId>
-  <artifactId>MinigamesProject</artifactId>
-  <version>1.19-SNAPSHOT</version>
-  <packaging>pom</packaging>
-  <name>Minigames-Project</name>
-  <url>https://www.spigotmc.org/resources/minigames.19687/</url>
-  <scm>
-    <connection>scm:git:git@github.com:AddstarMC/Minigames.git</connection>
-    <developerConnection>scm:git:git@github.com:AddstarMC/Minigames.git</developerConnection>
-    <url>https://github.com/AddstarMC/Minigames</url>
-    <tag>HEAD</tag>
-  </scm>
-  <properties>
-      <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-      <build.number />
-      <plugin.version>${project.version}-${build.number}</plugin.version>
-      <maven.javadoc.failOnError>false</maven.javadoc.failOnError>
-      <spigot.version>${spigot.api.version}.${spigot.minor.version}-R0.1-SNAPSHOT</spigot.version>
-      <spigot.api.version>1.19</spigot.api.version>
-      <spigot.minor.version>3</spigot.minor.version>
-      <codacy.api.token>${env.CODACY_API_TOKEN}</codacy.api.token>
-      <!--suppress UnresolvedMavenProperty -->
-      <travis.commit>${env.TRAVIS_COMMIT}</travis.commit>
-      <env.CODACY_API_TOKEN/>
-      <argline/>
-  </properties>
-  <developers>
-    <developer>
-      <id>Razz</id>
-      <name>Razz</name>
-      <url>http://github.com/Razz</url>
-      <roles>
-        <role>emeritus</role>
-      </roles>
-    </developer>
-    <developer>
-      <id>AddstarMC</id>
-      <name>AddstarMC</name>
-      <url>http://github.com/AddstarMC</url>
-      <roles>
-        <role>developer</role>
-        <role>maintainer</role>
-      </roles>
-      <timezone>10</timezone>
-    </developer>
-  </developers>
-  <licenses>
-    <license>
-        <name>MIT License</name>
-        <url>https://opensource.org/licenses/MIT</url>
-    </license>
-  </licenses>
-  <distributionManagement>
-    <repository>
-      <id>maven.addstar.com.au</id>
-      <name>maven.addstar.com.au-releases</name>
-      <url>https://maven.addstar.com.au/artifactory/ext-release-local</url>
-    </repository>
-      <snapshotRepository>
-      <id>maven.addstar.com.au</id>
-      <name>maven.addstar.com.au-snapshots</name>
-      <url>https://maven.addstar.com.au/artifactory/ext-snapshot-local</url>
-    </snapshotRepository>
-  </distributionManagement>
-  <modules>
-    <module>Minigames</module>
-    <module>Regions</module>
-    <module>minigamesdistribution</module>
-  </modules>
-  <repositories>
-    <!-- Addstar repo -->
-      <repository>
-          <id>addstar-repo</id>
-          <url>https://maven.addstar.com.au/artifactory/ext-release-local</url>
-          <snapshots>
-              <enabled>false</enabled>
-          </snapshots>
-      </repository>
-      <repository>
-          <id>addstar-snapshot-repo</id>
-          <url>https://maven.addstar.com.au/artifactory/ext-snapshot-local</url>
-          <snapshots>
-              <enabled>true</enabled>
-          </snapshots>
-          <releases>
-              <enabled>false</enabled>
-          </releases>
-      </repository>
-      <repository>
-          <id>papermc</id>
-          <url>https://repo.papermc.io/repository/maven-public/</url>
-      </repository>
-    <!-- Other repositories -->
-      <repository>
-          <id>spigot-repo</id>
-          <url>https://hub.spigotmc.org/nexus/content/groups/public/</url>
-      </repository>
-    <repository>
-      <id>maven-central</id>
-      <url>https://repo1.maven.org/maven2/</url>
-  </repository>
-      <repository>
-          <snapshots>
-              <enabled>false</enabled>
-          </snapshots>
-          <id>jcenter</id>
-          <name>bintray</name>
-          <url>https://jcenter.bintray.com</url>
-      </repository>
-  </repositories>
-  <dependencies>
-      <dependency>
-          <groupId>io.papermc.paper</groupId>
-          <artifactId>paper-api</artifactId>
-          <version>${spigot.version}</version>
-          <scope>provided</scope>
-      </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>4.13.1</version>
-      <scope>test</scope>
-    </dependency>
-      <dependency>
-          <groupId>org.jetbrains</groupId>
-          <artifactId>annotations</artifactId>
-          <version>20.1.0</version>
-          <scope>provided</scope>
-          <!-- no runtime features are used-->
-      </dependency>
-  </dependencies>
+<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <groupId>au.com.mineauz</groupId>
+    <artifactId>MinigamesProject</artifactId>
+    <version>1.19-SNAPSHOT</version>
+    <packaging>pom</packaging>
+    <name>Minigames-Project</name>
+    <url>https://www.spigotmc.org/resources/minigames.19687/</url>
+    <scm>
+        <connection>scm:git:git@github.com:AddstarMC/Minigames.git</connection>
+        <developerConnection>scm:git:git@github.com:AddstarMC/Minigames.git</developerConnection>
+        <url>https://github.com/AddstarMC/Minigames</url>
+        <tag>HEAD</tag>
+    </scm>
+    <properties>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <build.number/>
+        <plugin.version>${project.version}-${build.number}</plugin.version>
+        <maven.javadoc.failOnError>false</maven.javadoc.failOnError>
+        <spigot.version>${spigot.api.version}.${spigot.minor.version}-R0.1-SNAPSHOT</spigot.version>
+        <spigot.api.version>1.19</spigot.api.version>
+        <spigot.minor.version>3</spigot.minor.version>
+        <codacy.api.token>${env.CODACY_API_TOKEN}</codacy.api.token>
+        <!--suppress UnresolvedMavenProperty -->
+        <travis.commit>${env.TRAVIS_COMMIT}</travis.commit>
+        <env.CODACY_API_TOKEN/>
+        <argline/>
+    </properties>
+    <developers>
+        <developer>
+            <id>Razz</id>
+            <name>Razz</name>
+            <url>http://github.com/Razz</url>
+            <roles>
+                <role>emeritus</role>
+            </roles>
+        </developer>
+        <developer>
+            <id>AddstarMC</id>
+            <name>AddstarMC</name>
+            <url>http://github.com/AddstarMC</url>
+            <roles>
+                <role>developer</role>
+            </roles>
+            <timezone>10</timezone>
+        </developer>
+        <developer>
+            <id>GreenSurvivors</id>
+            <name>GreenSurvivors Team</name>
+            <url>http://github.com/GreenSurvivors</url>
+            <roles>
+                <role>developer</role>
+                <role>maintainer</role>
+            </roles>
+        </developer>
+    </developers>
+    <licenses>
+        <license>
+            <name>MIT License</name>
+            <url>https://opensource.org/licenses/MIT</url>
+        </license>
+    </licenses>
+    <distributionManagement>
+        <repository>
+            <id>maven.addstar.com.au</id>
+            <name>maven.addstar.com.au-releases</name>
+            <url>https://maven.addstar.com.au/artifactory/ext-release-local</url>
+        </repository>
+        <snapshotRepository>
+            <id>maven.addstar.com.au</id>
+            <name>maven.addstar.com.au-snapshots</name>
+            <url>https://maven.addstar.com.au/artifactory/ext-snapshot-local</url>
+        </snapshotRepository>
+    </distributionManagement>
+    <modules>
+        <module>Minigames</module>
+        <module>Regions</module>
+        <module>minigamesdistribution</module>
+    </modules>
+    <repositories>
+        <!-- Addstar repo -->
+        <repository>
+            <id>addstar-repo</id>
+            <url>https://maven.addstar.com.au/artifactory/ext-release-local</url>
+            <snapshots>
+                <enabled>false</enabled>
+            </snapshots>
+        </repository>
+        <repository>
+            <id>addstar-snapshot-repo</id>
+            <url>https://maven.addstar.com.au/artifactory/ext-snapshot-local</url>
+            <snapshots>
+                <enabled>true</enabled>
+            </snapshots>
+            <releases>
+                <enabled>false</enabled>
+            </releases>
+        </repository>
+        <repository>
+            <id>papermc</id>
+            <url>https://repo.papermc.io/repository/maven-public/</url>
+        </repository>
+        <!-- Other repositories -->
+        <repository>
+            <id>spigot-repo</id>
+            <url>https://hub.spigotmc.org/nexus/content/groups/public/</url>
+        </repository>
+        <repository>
+            <id>maven-central</id>
+            <url>https://repo1.maven.org/maven2/</url>
+        </repository>
+        <repository>
+            <snapshots>
+                <enabled>false</enabled>
+            </snapshots>
+            <id>jcenter</id>
+            <name>bintray</name>
+            <url>https://jcenter.bintray.com</url>
+        </repository>
+    </repositories>
+    <dependencies>
+        <dependency>
+            <groupId>io.papermc.paper</groupId>
+            <artifactId>paper-api</artifactId>
+            <version>${spigot.version}</version>
+            <scope>provided</scope>
+        </dependency>
+        <dependency>
+            <!-- updating to junit 5 breaks the tests (yes I had updated the Test files)
+            It's because of org.junit.jupiter.api.Test
+            -->
+            <groupId>junit</groupId>
+            <artifactId>junit</artifactId>
+            <version>4.13.2</version>
+            <scope>test</scope>
+        </dependency>
+        <!-- add just for the sake to stop the tests to whining about it missig
+        <dependency>
+        <groupId>org.slf4j</groupId>
+        <artifactId>slf4j-simple</artifactId>
+        <version>2.0.4</version>
+    </dependency> -->
+        <dependency>
+            <groupId>org.jetbrains</groupId>
+            <artifactId>annotations</artifactId>
+            <version>20.1.0</version>
+            <scope>provided</scope>
+            <!-- no runtime features are used-->
+        </dependency>
+    </dependencies>
 
-  <build>
-    <finalName>${project.artifactId}-${project.version}</finalName>
-    <pluginManagement>
-      <plugins>
-        <plugin>
-          <groupId>org.apache.maven.plugins</groupId>
-          <artifactId>maven-compiler-plugin</artifactId>
-          <version>3.8.1</version>
-          <configuration>
-            <source>1.8</source>
-            <target>1.8</target>
-          </configuration>
-        </plugin>
-        <plugin>
-          <artifactId>maven-assembly-plugin</artifactId>
-          <version>3.3.0</version>
-        </plugin>
-        <plugin>
-          <groupId>org.apache.maven.plugins</groupId>
-          <artifactId>maven-javadoc-plugin</artifactId>
-          <version>3.2.0</version>
-        </plugin>
-          <plugin>
-              <groupId>org.apache.maven.plugins</groupId>
-              <artifactId>maven-source-plugin</artifactId>
-              <version>3.2.1</version>
-              <executions>
-                  <execution>
-                      <id>attach-sources</id>
-                      <goals>
-                          <goal>jar</goal>
-                      </goals>
-                      <phase>package</phase>
-                  </execution>
-              </executions>
-              <configuration>
-                  <skipSource>false</skipSource>
-              </configuration>
-          </plugin>
-          <plugin>
-              <groupId>org.apache.maven.plugins</groupId>
-              <artifactId>maven-release-plugin</artifactId>
-              <version>2.5.3</version>
-          </plugin>
-          <plugin>
-              <groupId>org.apache.maven.plugins</groupId>
-              <artifactId>maven-dependency-plugin</artifactId>
-              <version>3.1.2</version>
-          </plugin>
-          <plugin>
-              <groupId>org.apache.maven.plugins</groupId>
-              <artifactId>maven-project-info-reports-plugin</artifactId>
-              <version>3.1.1</version>
-          </plugin>
-          <plugin>
-              <groupId>org.apache.maven.plugins</groupId>
-              <artifactId>maven-surefire-plugin</artifactId>
-              <version>2.22.2</version>
-              <configuration>
-                      <!-- no argLine here -->
-              </configuration>
-          </plugin>
-          <plugin>
-              <groupId>org.jacoco</groupId>
-              <artifactId>jacoco-maven-plugin</artifactId>
-              <version>0.8.5</version>
-              <executions>
-                  <execution>
-                      <id>prepare</id>
-                      <goals>
-                          <goal>prepare-agent</goal>
-                      </goals>
-                      <phase>compile</phase>
-                  </execution>
-              </executions>
-          </plugin>
-      </plugins>
-    </pluginManagement>
-  </build>
+    <build>
+        <finalName>${project.artifactId}-${project.version}</finalName>
+        <pluginManagement>
+            <plugins>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-compiler-plugin</artifactId>
+                    <version>3.11.0</version>
+                    <configuration>
+                        <source>17</source>
+                        <target>17</target>
+                    </configuration>
+                </plugin>
+                <plugin>
+                    <artifactId>maven-assembly-plugin</artifactId>
+                    <version>3.5.0</version>
+                </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-javadoc-plugin</artifactId>
+                    <version>3.5.0</version>
+                </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-source-plugin</artifactId>
+                    <version>3.2.1</version>
+                    <executions>
+                        <execution>
+                            <id>attach-sources</id>
+                            <goals>
+                                <goal>jar</goal>
+                            </goals>
+                            <phase>package</phase>
+                        </execution>
+                    </executions>
+                    <configuration>
+                        <skipSource>false</skipSource>
+                    </configuration>
+                </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-release-plugin</artifactId>
+                    <version>3.0.0</version>
+                </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-dependency-plugin</artifactId>
+                    <version>3.5.0</version>
+                </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-project-info-reports-plugin</artifactId>
+                    <version>3.1.1</version>
+                </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-surefire-plugin</artifactId>
+                    <version>3.0.0-M6</version>
+                    <configuration>
+                        <!-- no argLine here -->
+                    </configuration>
+                </plugin>
+                <plugin>
+                    <groupId>org.jacoco</groupId>
+                    <artifactId>jacoco-maven-plugin</artifactId>
+                    <version>0.8.5</version>
+                    <executions>
+                        <execution>
+                            <id>prepare</id>
+                            <goals>
+                                <goal>prepare-agent</goal>
+                            </goals>
+                            <phase>compile</phase>
+                        </execution>
+                    </executions>
+                </plugin>
+            </plugins>
+        </pluginManagement>
+    </build>
     <reporting>
         <plugins>
             <plugin>
                 <groupId>org.codehaus.mojo</groupId>
                 <artifactId>versions-maven-plugin</artifactId>
-                <version>2.8.1</version>
+                <version>2.15.0</version>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
