@@ -10,14 +10,19 @@ import javax.annotation.Nullable;
 import java.util.UUID;
 
 public class EntityData {
+    //uuid of the entity
     private final UUID uuid;
+    //type of entity
     private final EntityType entType;
+    //location the entity had when it was changed
     private final Location entLocation;
-    private final MinigamePlayer player;
+    // the player who has changed the entity. If null, the entity doesn't get reset if the player left the minigame
+    private final @Nullable MinigamePlayer player;
+    // was this entity created and needs to removed or was it changed / killed?
     private final boolean created;
 
     //todo save metadata like armor
-    public EntityData(Entity entity, MinigamePlayer modifier, boolean created) {
+    public EntityData(Entity entity, @Nullable MinigamePlayer modifier, boolean created) {
         this.uuid = entity.getUniqueId();
         this.entType = entity.getType();
         this.entLocation = entity.getLocation();
@@ -29,7 +34,7 @@ public class EntityData {
         return Bukkit.getEntity(uuid);
     }
 
-    public MinigamePlayer getModifier() {
+    public @Nullable MinigamePlayer getModifier() {
         return player;
     }
 
