@@ -1,4 +1,4 @@
 /**
- * BlockRecorder is the system for recording block changes in a game and rolling themback afterward.
+ * BlockRecorder is the system for recording block / entity changes in a game and rolling them back afterward.
  */
 package au.com.mineauz.minigames.blockRecorder;
