@@ -8,7 +8,7 @@ import au.com.mineauz.minigames.display.spigot.SpigotDisplayPoint;
 import au.com.mineauz.minigames.objects.MgRegion;
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.SetMultimap;
-import org.apache.commons.lang.Validate;
+import org.apache.commons.lang3.Validate;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.World;
