@@ -34,22 +34,22 @@ public class KillAction extends AbstractAction {
         return true;
     }
 
-		@Override
-		public void executeNodeAction(MinigamePlayer player,
-				Node node) {
-			debug(player,node);
-			if(player == null || !player.isInMinigame()) return;
-			if(!player.isLiving())
-				player.getPlayer().setHealth(0.0);
-		}
+    @Override
+    public void executeNodeAction(MinigamePlayer player,
+                                  Node node) {
+        debug(player, node);
+        if (player == null || !player.isInMinigame()) return;
+        if (!player.isLiving())
+            player.getPlayer().setHealth(0.0);
+    }
 
-		@Override
-		public void executeRegionAction(MinigamePlayer player, Region region) {
-			debug(player,region);
-			if(player == null || !player.isInMinigame()) return;
-			if(player.isLiving())
-				player.getPlayer().setHealth(0.0);
-		}
+    @Override
+    public void executeRegionAction(MinigamePlayer player, Region region) {
+        debug(player, region);
+        if (player == null || !player.isInMinigame()) return;
+        if (player.isLiving())
+            player.getPlayer().setHealth(0.0);
+    }
 
     @Override
     public void saveArguments(FileConfiguration config, String path) {
